:recycle: Remove transfer repository and move create transfer method into balance repository
