+++ src/repositories/__init__.py
@@ -5,5 +5,4 @@ from .secret_medias import *
 from .secret_messages import *
 from .team_members import *
 from .teams import *
-from .transfers import *
 from .users import *
